Fix auto-closing door config option, and make iron doors behave more like vanilla
Only queue iron door to be auto-closed if the config file says we should. Add a default `false` to the config since that's the current behavior.

Do not open iron doors if the player is sneaking AND has an item in their hand, as vanilla wooden doors would not open in this case.